Kraken: add structure for mock kraken
mock kraken will be kraken with already loaded data.

they'll be used in integration tests
