+++ source/kraken/CMakeLists.txt
@@ -1,5 +1,8 @@
-add_executable(kraken kraken_zmq.cpp worker.cpp maintenance_worker.cpp)
-target_link_libraries(kraken disruption_api calendar_api zmq time_tables types autocomplete proximitylist
+
+add_library(workers worker.cpp maintenance_worker.cpp)
+
+add_executable(kraken kraken_zmq.cpp)
+target_link_libraries(kraken workers disruption_api calendar_api zmq time_tables types autocomplete proximitylist
     ptreferential time_tables data routing fare georef utils SimpleAmqpClient
     rabbitmq-static log4cplus tcmalloc ${Boost_THREAD_LIBRARY}
     ${Boost_DATE_TIME_LIBRARY} ${Boost_SERIALIZATION_LIBRARY}
